+++ src/main/duke/task/ToDo.java
@@ -1,4 +1,4 @@
-package task;
+package main.duke.task;
 
 import java.util.HashMap;
